Add PagingWrapper instaed of legacy pager object

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/notification/ProgramNotificationInstanceService.java
@@ -45,7 +45,7 @@ public interface ProgramNotificationInstanceService
     List<ProgramNotificationInstance> getProgramNotificationInstances(
         ProgramNotificationInstanceParam programNotificationInstanceParam );
 
-    int countProgramNotificationInstances( ProgramNotificationInstanceParam params );
+    Long countProgramNotificationInstances( ProgramNotificationInstanceParam params );
 
     void validateQueryParameters( ProgramNotificationInstanceParam param );
 }
