+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/notification/ProgramNotificationInstanceStore.java
@@ -41,5 +41,5 @@ public interface ProgramNotificationInstanceStore
     List<ProgramNotificationInstance> getProgramNotificationInstances(
         ProgramNotificationInstanceParam programNotificationInstanceParam );
 
-    int countProgramNotificationInstances( ProgramNotificationInstanceParam params );
+    Long countProgramNotificationInstances( ProgramNotificationInstanceParam params );
 }
