+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/notification/DefaultProgramNotificationInstanceService.java
@@ -75,7 +75,7 @@ public class DefaultProgramNotificationInstanceService
     }
 
     @Override
-    public int countProgramNotificationInstances( ProgramNotificationInstanceParam params )
+    public Long countProgramNotificationInstances( ProgramNotificationInstanceParam params )
     {
         return notificationInstanceStore.countProgramNotificationInstances( params );
     }
